2/3 endpoints documentados
