+++ api/.phpunit.result.cache
@@ -1,1 +1,1 @@
-{"version":1,"defects":{"Tests\\Feature\\Users\\AuthTest::testLoginWithValidCredentials":7,"Tests\\Feature\\Users\\AuthTest::testLoginWithInvalidCredentials":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsers":5,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithValidFields":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithInvalidFields":7,"Tests\\Unit\\Users\\Auth\\MeTest::testGetInfoUserAuth":7,"Tests\\Unit\\Users\\Auth\\LogoutTest::testLogout":7,"Tests\\Unit\\Users\\Auth\\RefreshTest::testRefresh":7,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithValidFields":7,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":7,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":7,"Tests\\Feature\\Vacancies\\CreateTest::testCreateVacancies":7,"MyJobsTestSuite::testCreateUser":5,"MyJobsTestSuite::testLogin":5,"MyJobsTestSuite::testCreateVacancy":5,"MyJobsTestSuite::testRefreshToken":5,"MyJobsTestSuite::testMe":5,"MyJobsTestSuite::testLogout":5,"Tests\\Feature\\Vacancies\\EditTest::testEditVacancies":7,"Tests\\Feature\\Vacancies\\DeleteTest::testDeleteVacancies":7},"times":{"Tests\\Unit\\ExampleTest::test_that_true_is_true":0.01,"Tests\\Feature\\ExampleTest::test_the_application_returns_a_successful_response":0.02,"Tests\\Feature\\Users\\AuthTest::testLoginWithValidCredentials":0.162,"Tests\\Feature\\Users\\AuthTest::testLoginWithInvalidCredentials":0.019,"Tests\\Unit\\Users\\CreateTest::testCreateUsers":0.105,"Tests\\Feature\\Users\\LoginTest::testLoginWithValidCredentials":0.225,"Tests\\Feature\\Users\\LoginTest::testLoginWithInvalidCredentials":0.006,"Tests\\Unit\\Users\\LoginTest::testLoginWithValidCredentials":0.151,"Tests\\Unit\\Users\\LoginTest::testLoginWithInvalidCredentials":0.032,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":0.209,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":0.006,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithValidFields":0.028,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithInvalidFields":0.006,"Tests\\Unit\\Users\\Auth\\MeTest::testGetInfoUserAuth":0.02,"Tests\\Unit\\Users\\Auth\\LogoutTest::testLogout":0.02,"Tests\\Unit\\Users\\Auth\\RefreshTest::testRefresh":0.023,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithValidFields":0.086,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":0.024,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":0.017,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithInvalidFields":0.009,"Tests\\Feature\\Vacancies\\CreateTest::testCreateVacancies":0.026,"MyJobsTestSuite::testCreateUser":0.061,"MyJobsTestSuite::testLogin":0.001,"MyJobsTestSuite::testCreateVacancy":0.001,"MyJobsTestSuite::testRefreshToken":0.001,"MyJobsTestSuite::testMe":0.001,"MyJobsTestSuite::testLogout":0.001,"Tests\\Feature\\Vacancies\\SearchTest::testSearchVacancies":0.017,"Tests\\Feature\\Vacancies\\ShowTest::testShowVacancies":0.019,"Tests\\Feature\\Vacancies\\EditTest::testEditVacancies":0.018,"Tests\\Feature\\Vacancies\\DeleteTest::testDeleteVacancies":0.026}}
+{"version":1,"defects":{"Tests\\Feature\\Users\\AuthTest::testLoginWithValidCredentials":7,"Tests\\Feature\\Users\\AuthTest::testLoginWithInvalidCredentials":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsers":5,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithValidFields":7,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithInvalidFields":7,"Tests\\Unit\\Users\\Auth\\MeTest::testGetInfoUserAuth":7,"Tests\\Unit\\Users\\Auth\\LogoutTest::testLogout":7,"Tests\\Unit\\Users\\Auth\\RefreshTest::testRefresh":7,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithValidFields":7,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":7,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":7,"Tests\\Feature\\Vacancies\\CreateTest::testCreateVacancies":7,"MyJobsTestSuite::testCreateUser":5,"MyJobsTestSuite::testLogin":5,"MyJobsTestSuite::testCreateVacancy":5,"MyJobsTestSuite::testRefreshToken":5,"MyJobsTestSuite::testMe":5,"MyJobsTestSuite::testLogout":5,"Tests\\Feature\\Vacancies\\EditTest::testEditVacancies":7,"Tests\\Feature\\Vacancies\\DeleteTest::testDeleteVacancies":7},"times":{"Tests\\Unit\\ExampleTest::test_that_true_is_true":0.01,"Tests\\Feature\\ExampleTest::test_the_application_returns_a_successful_response":0.02,"Tests\\Feature\\Users\\AuthTest::testLoginWithValidCredentials":0.162,"Tests\\Feature\\Users\\AuthTest::testLoginWithInvalidCredentials":0.019,"Tests\\Unit\\Users\\CreateTest::testCreateUsers":0.105,"Tests\\Feature\\Users\\LoginTest::testLoginWithValidCredentials":0.225,"Tests\\Feature\\Users\\LoginTest::testLoginWithInvalidCredentials":0.006,"Tests\\Unit\\Users\\LoginTest::testLoginWithValidCredentials":0.151,"Tests\\Unit\\Users\\LoginTest::testLoginWithInvalidCredentials":0.032,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":0.209,"Tests\\Unit\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":0.006,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithValidFields":0.028,"Tests\\Unit\\Users\\CreateTest::testCreateUsersWithInvalidFields":0.006,"Tests\\Unit\\Users\\Auth\\MeTest::testGetInfoUserAuth":0.033,"Tests\\Unit\\Users\\Auth\\LogoutTest::testLogout":0.044,"Tests\\Unit\\Users\\Auth\\RefreshTest::testRefresh":0.033,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithValidFields":0.697,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithValidCredentials":0.012,"Tests\\Feature\\Auth\\Users\\LoginTest::testLoginWithInvalidCredentials":0.018,"Tests\\Feature\\Users\\CreateTest::testCreateUsersWithInvalidFields":0.071,"Tests\\Feature\\Vacancies\\CreateTest::testCreateVacancies":0.025,"MyJobsTestSuite::testCreateUser":0.061,"MyJobsTestSuite::testLogin":0.001,"MyJobsTestSuite::testCreateVacancy":0.001,"MyJobsTestSuite::testRefreshToken":0.001,"MyJobsTestSuite::testMe":0.001,"MyJobsTestSuite::testLogout":0.001,"Tests\\Feature\\Vacancies\\SearchTest::testSearchVacancies":0.061,"Tests\\Feature\\Vacancies\\ShowTest::testShowVacancies":0.058,"Tests\\Feature\\Vacancies\\EditTest::testEditVacancies":0.031,"Tests\\Feature\\Vacancies\\DeleteTest::testDeleteVacancies":0.033}}
